[libcxx][subobject-bounds] Add opt-out annotation to short string __data_ array
Without this annotation we were setting bounds of 1 when taking a reference
to an element in the array and converting that reference to a pointer.
This happened whenever __get_short_pointer() was used:

pointer __get_short_pointer() _NOEXCEPT {
  return pointer_traits<pointer>::pointer_to(__r_.first().__s.__data_[0]);
}

In this call `__r_.first().__s.__data_[0]` is turned into a C++ reference
(bounded to 1 element since we assume that references to an array element
should not span the whole array) and then passed to pointer_to.
The pointer_to call receives a char& bounded to 1 byte so it will return a
char* also bounded to 1 byte which then causes a bounds violation later on.

It would be nice if we could just annotate pointer_to and std::addressof()
in a way that all array element references passed to it are bounded to the
full array but this is not possible in the current clang implementation of
subobject-bounds.