[libcxx][subobject-bounds] Don't use address-of on std::string::operator[]
The libc++ locale header contained the following pattern multiple times:
To get a pointer the std::string} internal character array it would use
&__buf[0]. However, std::string has an operator[] that returns a char& and
should return a reference to only that element when subobject-bounds are
enabled. The fix for this is to use std::string::data() instead since this
returns a pointer to the underlying storage.

TODO: add compiler diagnostic for this pattern.