[UPSTREAM][libunwind][libcxx][libcxxabi][CMake] Share standalone build logic
Now that we require the monorepo layout we can use llvm-lit from the source
tree rather than having to pass a path explicitly for standalone builds.
This commit merges the two similar copies of HandleOutOfTreeLLVM and moves
them to llvm/cmake/modules/HandleStandaloneBuild.cmake.

This is mostly NFC, but it does remove the ability to query llvm-config for
the paths in libunwind/libcxxabi.