[libcxx][lit] Add a new executor to collect test binaries
Summary:
I originally added this executor for the old test format to the CHERI fork
of libc++ to gather all test binaries in one directory. This change is a
simple port of the executor to the new format.

This is mostly useful for debugging and performing analysis on the compiled
binaries or copying them in bulk to a remote server. Using this executor
also helped me discover the bug fixed in https://reviews.llvm.org/D84040.

Reviewers: #libc!, ldionne

Subscribers: dexonsmith, libcxx-commits

Tags: #libc

Differential Revision: https://reviews.llvm.org/D84045