[UPSTREAM][libunwind][libc++][libc++abi] Allow overriding the executor for tests
This is useful when trying to run multiple tests with different
arguments to the executor script. This is needed in my case since I
do not know the correct ssh connection arguments when building libc++
since the target port is only known when running the tests but not
when calling CMake. This change allows me to pass
`--param executor=ssh.py <args>` to lit once I know the right
hostname/port for running the tests.