[UPSTREAM] Use if constexpr to silence compiler warnings
This only matters downstream - in the CHERI target we specialize this
function and calling it from the specialized version triggers shift count
warnings in the else case. These warnings appear to not trigger when called
directly. Fix this by using if constexpr which is safe since the functions
are C++20 only.